Fixed error in reeceive_handler and minor additions
Added the failed_connection flag to DyrBot.
Added sleep_for(seconds(5)) to BotManager.
Fixed recbuf.pop() not being called for every call to receive_handler.
Fixed declaration of same variable twice, once as static and once as non-static in receive_handler.

Fixed order, so that ConnectionManager::process() is called in advanced of creating bots, so that request_to_connect can be executed as asoon as posisble.